Seperated MapServiceException from MapControllerService. Poi is now partially selected when navigating from tree page to map.
