+++ lib/services/map_controller_service.dart
@@ -1,7 +1,12 @@
+// Flutter
 import 'package:flutter/material.dart';
 
+// pub.dev
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
+
+// httapp
+import 'package:httapp/exceptions/map_service_exception.dart';
 
 class MapControllerService {
   static final MapControllerService _singleton = MapControllerService._internal();
@@ -10,9 +15,11 @@ class MapControllerService {
 
   late final MapController mapController;
   bool _initialized = false;
+  String? selectedPoiId;
   
   /// Initialize the service
   /// 
+  /// creates a new MapController if init has never been called before
   void init() {
     if (_initialized) return;
 
@@ -22,9 +29,29 @@ class MapControllerService {
     _initialized = true;
   }
 
+  /// Set the [selectedPoiId] to the provided id value.
+  void selectPoi(String poiId) {
+    selectedPoiId = poiId;
+  }
+
+  /// Deselects the current [selectedPoiId] to null
+  void deselectPoi() {
+    selectedPoiId = null;
+  }
+
   // moveAndRotate(LatLng center, double zoom, double degree, {String? id}) → MoveAndRotateResult 
   void moveAndRotate(LatLng center, double zoom, double degree) {
     try {
+      // check if init has been called,
+      // indicates if a MapController exists
+      if (!_initialized) {
+        throw MapServiceException(
+          method: 'moveAndRotate',
+          result: 'uninitialized',
+          message: 'Please call init() to start the service before attempting to use any methods.'
+        );
+      }
+
       // attempt to rotate and/or move the map
       var result = mapController.moveAndRotate(center, zoom, degree);
 
@@ -43,32 +70,5 @@ class MapControllerService {
     } catch(e) {
       debugPrint('[MapControllerService] $e');
     }
-  }
-}
-
-/// Exception thrown when the MapControllerService fails
-class MapServiceException implements Exception {
-  /// A message describing the exception 
-  final String message;
-
-  /// The method the exception is for
-  final String method;
-
-  /// The method's result that created the exception
-  final String result;
-
-  MapServiceException({
-    required this.method,
-    required this.result,
-    this.message = '',
-    });
-
-
-
-    @override
-  String toString() {
-    String output = '[$method] -> $result\n\t $message';
-
-    return output;
   }
 }
